budget/requirements.txt
asgiref==3.4.1
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
pika==1.1.0
pytz==2021.3
sqlparse==0.4.2
